Remove a pointer->integer warning in fstrace.c
Explicitly cast eparm to a long in afs_syscall when calling
proc_afs_syscall, to avoid a warning.

Reviewed-on: http://gerrit.openafs.org/596
Tested-by: Derrick Brashear <shadow@dementia.org>

+++ src/venus/fstrace.c
@@ -1135,7 +1135,7 @@ afs_syscall(long call, long parm0, long parm1, long parm2, long parm3,
 #endif
     /* Linux can only handle 5 arguments in the actual syscall. */
     if (call == AFSCALL_ICL) {
-	rval = proc_afs_syscall(call, parm0, parm1, parm2, eparm, &code);
+	rval = proc_afs_syscall(call, parm0, parm1, parm2, (long)eparm, &code);
 	if (rval)
 	    code = syscall(AFS_SYSCALL, call, parm0, parm1, parm2, eparm);
     } else {
